Fix Wall Werrors on modern (#3412)
* Fix Wall Werror errors

* fix hgss pokedex modern errors

* tests build on modern

* learnset change

* Update src/battle_ai_util.c


* Update src/battle_anim_new.c


* Update src/debug.c


* Update src/debug.c


* Update src/pokedex_plus_hgss.c


* review changes

* unused cmd args

* various args back

---------

+++ src/field_specials.c
@@ -964,8 +964,8 @@ void FieldShowRegionMap(void)
 
 static bool8 IsPlayerInFrontOfPC(void)
 {
-    u16 x, y;
-    u16 tileInFront;
+    s16 x, y;
+    u32 tileInFront;
 
     GetXYCoordsOneStepInFrontOfPlayer(&x, &y);
     tileInFront = MapGridGetMetatileIdAt(x, y);
